hacks/genai-multimodal/artifacts/function/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.84.0
google-cloud-bigquery==3.30.0
requests==2.32.3